feat: "Final adjustment in exercise 2"
